Data-Server/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.36.0
python-dotenv>=1.0.0
flasgger>=0.9.7
pandas>=2.0.0
numpy>=1.24.0
flask-cors>=4.0.0